#398 refactor OeZoneerderZonePanel component
