+++ src/components/smart/OeZoneerderZonePanel.vue
@@ -124,20 +124,17 @@
 import 'ol/ol.css';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { VlButton, VlInputField, VlInputGroup, VlLink, VlTitle } from '@govflanders/vl-ui-design-system-vue3';
-import { Feature, Map, MapBrowserEvent } from 'ol';
+import { Map, MapBrowserEvent } from 'ol';
 import { unByKey } from 'ol/Observable';
 import { type Extent } from 'ol/extent';
 import { GeoJSON, WKT } from 'ol/format';
-import { Circle, Geometry, MultiPolygon, Polygon } from 'ol/geom';
-import { fromCircle } from 'ol/geom/Polygon';
+import { Geometry } from 'ol/geom';
 import { Draw } from 'ol/interaction';
 import VectorLayer from 'ol/layer/Vector';
 import VectorSource from 'ol/source/Vector';
-import { Fill, Text as OlText, Stroke, Style } from 'ol/style';
 import { inject, onMounted, onUnmounted, ref, watch } from 'vue';
 import { FeatureSelectEnum } from '@models/featureSelect.enum';
-import { FeatureSelectConfig } from '@models/oe-zoneerder-config';
-import type { FeatureLike } from 'ol/Feature';
+import { FeatureSelectConfig } from '@models/oe-map-config';
 import type { ColorLike } from 'ol/colorlike';
 import type { CrabApiService } from '@/services';
 import type { Contour, IDrawGeomType } from '@models/oe-openlayers';
@@ -155,8 +152,14 @@ const elementRef = ref<HTMLElement>();
 const emit = defineEmits(['update:zone', 'update:feature-select', 'zone-panel:mounted']);
 
 const map = inject('map') as Map;
-const zoomToExtent = inject('zoomToExtent') as (extent: Extent) => void;
+const drawLayer = inject('drawLayer') as VectorLayer<VectorSource<Geometry>>;
 const crabService = inject('crabService') as CrabApiService;
+const zoomToExtent = inject('zoomToExtent') as (extent: Extent) => void;
+const createVectorLayer = inject('createVectorLayer') as (options: {
+  color: ColorLike;
+  fill: ColorLike;
+  title: string;
+}) => VectorLayer<VectorSource<Geometry>>;
 
 const WKTString = ref('');
 // GEOMETRYCOLLECTION(POLYGON ((161224.97845101 212135.39413324, 161227.78030702 212146.77013324, 161227.79144301 212149.23784525, 161227.18613102 212151.84712525, 161227.10414701 212155.79176525, 161227.01077101 212160.28341325, 161226.58331501 212161.16456525, 161219.72001901 212172.48002126, 161216.50203501 212180.24232526, 161215.619987 212182.37000527, 161210.879955 212192.90005327, 161206.650003 212200.76994128, 161204.026451 212205.67893328, 161202.14997099 212209.18997329, 161196.13006699 212222.70997329, 161195.37025899 212224.6055893, 161192.23003499 212232.4400213, 161188.71995499 212238.6398933, 161183.70997098 212247.58997331, 161182.12635498 212250.23675732, 161178.01998698 212257.09998932, 161174.55829097 212262.74952532, 161173.01998697 212265.26005333, 161171.94997098 212268.24002133, 161171.87995498 212268.61000533, 161174.68001898 212277.42402133, 161179.16155498 212296.29506135, 161179.09979498 212299.29941335, 161177.44718698 212299.33333335, 161120.17633894 212265.64232532, 161117.56142694 212260.76846932, 161116.00033893 212255.48296532, 161114.33013093 212251.16654931, 161094.47560292 212215.89243729, 161126.20827495 212188.88558927, 161134.88161895 212185.71592527, 161155.68385896 212173.54862926, 161159.50958697 212171.90786126, 161161.15457897 212170.49544526, 161163.44955497 212169.01435726, 161164.32494697 212165.10280526, 161164.48328297 212162.31176525, 161163.84174697 212157.13877325, 161164.75221097 212155.95643725, 161164.93704297 212153.82907725, 161162.30293097 212150.97518925, 161162.02721897 212148.16750924, 161163.17115497 212146.32686924, 161162.07003497 212142.38914124, 161214.410771 212117.61026122, 161216.702355 212116.52526922, 161224.02869101 212113.05698122, 161224.97845101 212135.39413324)))
@@ -174,18 +177,8 @@ const geoJsonFormatter = new GeoJSON({
   dataProjection: mapProjection,
   featureProjection: mapProjection,
 });
-const drawLayer = _createVectorLayer({
-  color: 'rgb(39, 146, 195)',
-  fill: 'rgba(39, 146, 195, 0.3)',
-  title: 'Zone',
-});
-drawLayer.getSource()?.on('addfeature', () => {
-  drawLayerToZone();
-});
-map.addLayer(drawLayer);
-addZoneToDrawLayer();
 
-const flashLayer = _createVectorLayer({
+const flashLayer = createVectorLayer({
   color: 'rgba(255,0,255, 1)',
   fill: 'rgba(255,0,255, 0.4)',
   title: '',
@@ -206,6 +199,12 @@ watch(
 
 onMounted(() => {
   emit('zone-panel:mounted', elementRef.value);
+  drawLayer
+    .getSource()
+    ?.getFeatures()
+    .forEach((feature) => {
+      geometryObjectList.value.push(feature.get('name'));
+    });
 });
 
 onUnmounted(() => {
@@ -214,6 +213,7 @@ onUnmounted(() => {
     .getArray()
     .filter((control) => control['element'].classList.contains('zone-panel'))
     .forEach((control) => map.removeControl(control));
+  geometryObjectList.value = [];
 });
 
 const featureSelectCallback = (
@@ -249,32 +249,6 @@ function togglePanel() {
 function updateWKTString(value: string) {
   invalidWKT.value = false;
   WKTString.value = value;
-}
-
-function _createVectorLayer(options: { color: ColorLike; fill: ColorLike; title: string }) {
-  const getText = (feature: FeatureLike) =>
-    new OlText({
-      font: '10px Verdana',
-      text: feature.get('name') || '',
-      fill: new Fill({ color: options.color }),
-      stroke: new Stroke({ color: '#fff', width: 3 }),
-    });
-
-  const getStyle = (feature: FeatureLike) =>
-    new Style({
-      stroke: new Stroke({ color: options.color, width: 3 }),
-      fill: new Fill({ color: options.fill }),
-      text: getText(feature),
-    });
-  const vLayer = new VectorLayer({
-    source: new VectorSource(),
-    style: getStyle,
-    visible: true,
-  });
-  vLayer.set('title', options.title);
-  vLayer.set('type', 'overlay');
-
-  return vLayer;
 }
 
 function toggleDrawZone(drawZoneEnabled = false, type: IDrawGeomType = 'Polygon') {
@@ -394,69 +368,8 @@ function removeGeometryObject(name: string) {
   featuresToRemove?.forEach((featureToRemove) => {
     drawLayerSource?.removeFeature(featureToRemove);
   });
-  drawLayerToZone();
-  if (zone.value?.coordinates.length === 0) {
-    zone.value = undefined;
-  }
 
   geometryObjectList.value.splice(geometryObjectList.value.indexOf(name), 1);
-}
-
-function drawLayerToZone() {
-  const multiPolygon = new MultiPolygon([], 'XY');
-  const features = drawLayer.getSource()?.getFeatures();
-  features?.forEach((feature) => {
-    const geom = feature.getGeometry();
-    if (geom instanceof Polygon) {
-      multiPolygon.appendPolygon(geom as Polygon);
-    } else if (geom instanceof MultiPolygon) {
-      geom.getPolygons().forEach((polygon: Polygon) => {
-        multiPolygon.appendPolygon(polygon);
-      });
-    } else if (geom instanceof Circle) {
-      multiPolygon.appendPolygon(fromCircle(geom));
-    }
-  });
-
-  const contour = formatGeoJson(multiPolygon);
-  if (!zone.value) {
-    zone.value = contour;
-  } else {
-    zone.value.coordinates = contour.coordinates;
-  }
-}
-
-function formatGeoJson(feature: Geometry) {
-  const geojson = geoJsonFormatter.writeGeometryObject(feature) as Contour;
-  // hack to add crs. todo: remove when https://github.com/openlayers/ol3/issues/2078 is fixed
-  geojson.crs = {
-    type: 'name',
-    properties: {
-      name: 'urn:ogc:def:crs:EPSG::31370',
-    },
-  };
-  return geojson as Contour;
-}
-
-function addZoneToDrawLayer() {
-  if (!drawLayer) return;
-
-  const drawSource = drawLayer.getSource() as VectorSource<Geometry>;
-  drawSource.getFeatures().forEach((feature) => {
-    drawSource.removeFeature(feature);
-  });
-  geometryObjectList.value = [];
-
-  if (!zone.value) return;
-  const coordinates = zone.value.type === 'Polygon' ? [zone.value.coordinates] : zone.value.coordinates;
-
-  for (const [index, coords] of coordinates.entries()) {
-    const name = coordinates.length > 1 ? `Zone feature ${index + 1}` : 'Zone';
-    const geometry = new Polygon(coords);
-    drawSource.addFeature(new Feature({ name, geometry }));
-    geometryObjectList.value.push(name);
-  }
-  zoomToExtent(geoJsonFormatter.readGeometry(zone.value).getExtent());
 }
 
 function flashFeature(featureName: string) {
