+++ src/components/smart/index.ts
@@ -6,6 +6,7 @@ import OeActorWidget from './OeActorWidget.vue';
 import OeActorWidgetGrid from './OeActorWidgetGrid.vue';
 import OeAdres from './OeAdres.vue';
 import OeInventarisLink from './OeInventarisLink.vue';
+import OeMap from './OeMap.vue';
 import OeReferences from './OeReferences.vue';
 import OeZoneerder from './OeZoneerder.vue';
 
@@ -20,4 +21,5 @@ export {
   OeReferences,
   OeZoneerder,
   OeInventarisLink,
+  OeMap,
 };
